rm duplicated class: bert_tokenizer - was in profile_corpus.py and process.py - moved to new bert_tokenier.py - renamed to BertTokenizerCustom
